requirements.txt
auraloss==0.4.0
cdpam==0.0.6
decord==0.6.0
einops==0.8.1
imageio==2.33
julius==0.2.7
librosa==0.9.2
matplotlib==3.7.2
numpy==1.23.5
Pillow==11.1.0
plyfile==1.1
pypesq==1.2.4
python_sofa==0.2.0
PyYAML==6.0.2
scikit_learn==1.3.0
scipy==1.15.2
scikit-image==0.23.2
soundfile==0.12.1
tensorboardX==2.6.2.2
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
yacs==0.1.8